+++ viceemulatorcore.cpp
@@ -90,7 +90,7 @@ void ViceEmulatorCore::RunMainVice(bool wait) {
 
   printf("Starting emulator main loop\n");
 
-#if defined(RASPI_C64) || defined(RASPI_C128)
+#if defined(RASPI_C64)
   int argc = 11;
   char *argv[] = {
       (char *)"vice", timing_option_, (char *)"-sounddev", (char *)"raspi",
@@ -98,6 +98,16 @@ void ViceEmulatorCore::RunMainVice(bool wait) {
       (char *)"-refresh", (char *)"1",
       // Unless we disable the video cache, vsync is messed up
       (char *)"+VICIIvcache",
+  };
+#elif defined(RASPI_C128)
+  int argc = 12;
+  char *argv[] = {
+      (char *)"vice", timing_option_, (char *)"-sounddev", (char *)"raspi",
+      (char *)"-soundoutput", (char *)"1", (char *)"-soundsync", (char *)"0",
+      (char *)"-refresh", (char *)"1",
+      // Unless we disable the video cache, vsync is messed up
+      (char *)"+VICIIvcache",
+      (char *)"+VDCvcache",
   };
 #elif defined(RASPI_VIC20)
   int argc = 11;